Figure out class code page and make minor adjustments to props of Answer questions class
